Improve website functionality, accessibility, and code structure
This commit includes several improvements to the website:

- **Bug Fix:** The award badges in the projects section were not clickable. This has been fixed by adding the `data-award-url` attribute to the anchor tags.
- **Code Refactoring:** The main `script.js` file has been split into smaller, more manageable modules to improve code organization and maintainability.
- **Accessibility:**
    - Added `aria-label` attributes to icon-only buttons for better screen reader support.
    - Improved color contrast for some elements to meet accessibility standards.
- **Cleanup:** Removed unused CSS rules.
