+++ index.html
@@ -105,11 +105,11 @@
                         </article>
                         <article class="project-card">
                             <div class="awards-container">
-                                <a href="https://t.me/m4xter/81" target="_blank" rel="noopener noreferrer" class="award-badge">
+                                <a href="https://t.me/m4xter/81" target="_blank" rel="noopener noreferrer" class="award-badge" data-award-url="https://t.me/m4xter/81">
                                     <span class="material-icons-outlined" aria-hidden="true">emoji_events</span>
                                     <span class="translate-element" data-key="judges-5th">5 место (судьи)</span>
                                 </a>
-                                <a href="https://itch.io/jam/maxters-jam-2024/results?page=2#:~:text=%D0%BE%D1%82%20ineedmypills-,28th%20%D0%BC%D0%B5%D1%81%D1%82%D0%BE,-%D1%81%2014%20%D0%BE%D1%86%D0%B5%D0%BD%D0%BA%D0%B0%D0%BC%D0%B8" target="_blank" rel="noopener noreferrer" class="award-badge">
+                                <a href="https://itch.io/jam/maxters-jam-2024/results?page=2#:~:text=%D0%BE%D1%82%20ineedmypills-,28th%20%D0%BC%D0%B5%D1%81%D1%82%D0%BE,-%D1%81%2014%20%D0%BE%D1%86%D0%B5%D0%BD%D0%BA%D0%B0%D0%BC%D0%B8" target="_blank" rel="noopener noreferrer" class="award-badge" data-award-url="https://itch.io/jam/maxters-jam-2024/results?page=2#:~:text=%D0%BE%D1%82%20ineedmypills-,28th%20%D0%BC%D0%B5%D1%81%D1%82%D0%BE,-%D1%81%2014%20%D0%BE%D1%86%D0%B5%D0%BD%D0%BA%D0%B0%D0%BC%D0%B8">
                                     <span class="material-icons-outlined" aria-hidden="true">star_border</span>
                                     <span class="translate-element" data-key="players-28th">28 место (игроки)</span>
                                 </a>
@@ -364,7 +364,7 @@
                         <p class="support-category-desc translate-element" data-key="crypto-desc">Для ценителей анонимности и современных технологий</p>
                     </div>
                     <div class="support-list">
-                        <button type="button" class="support-item support-crypto-ton" data-address="UQALAY9I3rkVFoJKZRSMH2ikuQZkcHPV318xZMiLiLtCPsi8">
+                        <button type="button" class="support-item support-crypto-ton" data-address="UQALAY9I3rkVFoJKZRSMH2ikuQZkcHPV318xZMiLiLtCPsi8" aria-label="Copy TON Space Address">
                             <div class="support-icon">
                                 <div class="icon-bg"></div>
                                 <svg viewBox="12 14 32 32" aria-hidden="true"><path d="M37.5603 15.6277H18.4386C14.9228 15.6277 12.6944 19.4202 14.4632 22.4861L26.2644 42.9409C27.0345 44.2765 28.9644 44.2765 29.7345 42.9409L41.5381 22.4861C43.3045 19.4251 41.0761 15.6277 37.5627 15.6277H37.5603ZM26.2548 36.8068L23.6847 31.8327L17.4833 20.7414C17.0742 20.0315 17.5795 19.1218 18.4362 19.1218H26.2524V36.8092L26.2548 36.8068ZM38.5108 20.739L32.3118 31.8351L29.7417 36.8068V19.1194H37.5579C38.4146 19.1194 38.9199 20.0291 38.5108 20.739Z" /></svg>
@@ -378,7 +378,7 @@
                                 <span class="translate-element" data-key="optimal">Оптимально</span>
                             </div>
                         </button>
-                        <button type="button" class="support-item support-crypto-usdt" data-address="TGFT2Y9cwB5ipbScFmWqmAAwEyLU754PUL">
+                        <button type="button" class="support-item support-crypto-usdt" data-address="TGFT2Y9cwB5ipbScFmWqmAAwEyLU754PUL" aria-label="Copy USDT (TRC20) Address">
                             <div class="support-icon">
                                 <div class="icon-bg"></div>
                                 <svg viewBox="0 0 112 90" aria-hidden="true">
@@ -467,7 +467,7 @@
     </div>
 </div>
 
-<script src="script.js"></script>
+<script type="module" src="js/main.js"></script>
 
 </body>
 </html>
