+++ js/state.js
@@ -1,0 +1,27 @@
+export const state = {
+    lang: 'ru',
+    theme: 'light',
+};
+
+export const BIRTH_DATE = '2006-06-12';
+
+export const selectors = {
+    themeToggle: '#themeToggle',
+    agePlaceholder: '#agePlaceholder',
+    shareButton: '#shareButton',
+    shareModal: '#shareModal',
+    shareModalClose: '#shareModalCloseButton',
+    shareOptions: '#shareOptions',
+    shareLinkInput: '#shareLinkInput',
+    copyLinkButton: '#copyLinkButton',
+    headerOptionsSubgroup: '#header-options-subgroup',
+    youtuberButton: '#youtuber-button',
+    gamedevInterestBadge: '#gamedev-interest-badge',
+    softwareInterestBadge: '#software-interest-badge',
+    translatableElement: '[data-key]',
+    animateOnScroll: '.animate-on-scroll',
+    staggerChild: '.stagger-child',
+    cryptoButton: '[data-address]',
+    awardBadge: '[data-award-url]',
+    contentSection: 'main.content > section[id]',
+};
