+++ js/translations.js
@@ -1,0 +1,142 @@
+export const translations = {
+    ru: {
+        "page-title": "ineedmypills",
+        "main-title": "ineedmypills",
+        "alt-header-title": "Коктейльчик",
+        "full-name": "Глухих Егор Алексеевич",
+        "specialist": "Начинающий многосторонне развитый IT-\"специалист\"",
+        "skills-title": "Технические навыки",
+        "lang-title": "Языки программирования",
+        "software-title": "Программное обеспечение",
+        "ai-skill": "Владение ИИ",
+        "projects-title": "Ключевые проекты",
+        "games-category-title": "Игры",
+        "software-category-title": "Программное обеспечение",
+        "fh3-desc": "Неофициальное переиздание игры Final Hours 3",
+        "npc-desc": "\"Почувствуйте, каково это - быть неигровым персонажем.\"",
+        "npc-jam": "Игра для",
+        "judges-5th": "5 место (судьи)",
+        "players-28th": "28 место (игроки)",
+        "kolobok-desc": "Увлекательный 3D шутер-платформер, вдохновлённый игрой \"My Friend Pedro\", с видом сбоку в стилистике \"Киберпанка\"",
+        "obs-desc": "Компактный экранный индикатор для записи и буфера повтора в OBS Studio",
+        "completed": "Завершенный проект",
+        "in-progress": "В разработке",
+        "course-completed": "Окончил",
+        "course-in-progress": "В процессе",
+        "education-title": "Образование",
+        "secondary-edu": "Среднее общее образование",
+        "11-classes": "11 классов",
+        "computer-academy": "Компьютерная Академия \"ТОП\"",
+        "specialization": "Специализация: разработка программного обеспечения",
+        "first-year": "1 курс",
+        "second-year": "2 курс",
+        "third-year": "3 курс",
+        "first-place": "1 место в группе",
+        "project-defense": "Защитил проект по Unreal Engine",
+        "contacts-title": "Контакты",
+        "about-title": "О себе",
+        "about-text": "Увлеченный геймдевелопер с разносторонним техническим бэкграундом. Страстно увлечен созданием игр - от концепции до релиза. Участвую в Game Jams для оттачивания навыков работы в сжатые сроки. Экспериментирую с интеграцией ИИ в творческие процессы. Обладаю упорством в доведении проектов до завершения и креативным подходом к решению технических задач. Постоянно осваиваю новые технологии и инструменты разработки.",
+        "interest-gamedev": "Разработка инди-игр",
+        "interest-software": "Разработка ПО",
+        "interest-ai": "Генеративное искусство и нейросети",
+        "interest-music": "Создание электронной музыки",
+        "interest-youtuber": "Ютубер",
+        "support-title": "Поддержать",
+        "t-bank": "Т-Банк",
+        "no-commission": "Без комиссии",
+        "commission-3.5": "Комиссия 3,5%",
+        "commission-10": "Комиссия 10%",
+        "recommended": "Рекомендуется",
+        "not-recommended-rf": "не рекомендуется для РФ",
+        "backup-option": "Запасной вариант",
+        "bank-transfer-title": "Банковский перевод",
+        "donations-title": "Донаты",
+        "crypto-title": "Криптовалюта",
+        "bank-desc": "Самый удобный способ без комиссии",
+        "donations-desc": "Если хотите, чтобы ваше сообщение появилось на стриме",
+        "crypto-desc": "Для ценителей анонимности и современных технологий",
+        "copy-to-clipboard": "Нажмите, чтобы скопировать",
+        "copied-feedback": "Скопировано!",
+        "share-title": "Конструктор ссылки",
+        "share-desc": "Выберите секции, которые хотите показать:",
+        "share-copy": "Копировать",
+        "share-copied": "Скопировано!",
+        "share-button-aria": "Поделиться",
+        "share-header": "Заголовок",
+        "share-alt-header": "Альтернативный заголовок",
+        "share-subtitle": "Подзаголовок",
+        "share-full-name": "ФИО",
+        "optimal": "Оптимально",
+        "lts": "Долгосрочная поддержка"
+    },
+    en: {
+        "page-title": "ineedmypills",
+        "main-title": "ineedmypills",
+        "alt-header-title": "Milky",
+        "full-name": "Glukhikh Egor Alekseevich",
+        "specialist": "Beginning versatile IT \"specialist\"",
+        "skills-title": "Technical Skills",
+        "lang-title": "Programming Languages",
+        "software-title": "Software",
+        "ai-skill": "AI Proficiency",
+        "projects-title": "Key Projects",
+        "games-category-title": "Games",
+        "software-category-title": "Software",
+        "fh3-desc": "Unofficial remake of Final Hours 3 game",
+        "npc-desc": "\"Experience what it's like to be a non-playable character.\"",
+        "npc-jam": "Game for",
+        "judges-5th": "5th place (judges)",
+        "players-28th": "28th place (players)",
+        "kolobok-desc": "Engaging 3D shooter-platformer inspired by \"My Friend Pedro\", with a side view in \"Cyberpunk\" style",
+        "obs-desc": "Compact on-screen indicator for OBS Studio recording and replay buffer status",
+        "completed": "Completed project",
+        "in-progress": "In development",
+        "course-completed": "Finished",
+        "course-in-progress": "In progress",
+        "education-title": "Education",
+        "secondary-edu": "General secondary education",
+        "11-classes": "11 grades",
+        "computer-academy": "TOP Computer Academy",
+        "specialization": "Specialization: software development",
+        "first-year": "1st course",
+        "second-year": "2nd course",
+        "third-year": "3rd course",
+        "first-place": "1st place in the group",
+        "project-defense": "Defended Unreal Engine project",
+        "contacts-title": "Contacts",
+        "about-title": "About Me",
+        "about-text": "Passionate game developer with a versatile technical background. Deeply engaged in game creation from concept to release. Participate in Game Jams to hone skills in tight deadlines. Experiment with AI integration in creative processes. Possess persistence in completing projects and a creative approach to solving technical challenges. Continuously learning new technologies and development tools.",
+        "interest-gamedev": "Indie Game Development",
+        "interest-software": "Software Development",
+        "interest-ai": "Generative Art and Neural Networks",
+        "interest-music": "Electronic Music Creation",
+        "interest-youtuber": "YouTuber",
+        "support-title": "Support",
+        "t-bank": "T-Bank",
+        "no-commission": "No commission",
+        "commission-3.5": "Commission 3.5%",
+        "commission-10": "Commission 10%",
+        "recommended": "Recommended",
+        "not-recommended-rf": "not recommended for RF",
+        "backup-option": "Backup option",
+        "bank-transfer-title": "Bank Transfer",
+        "donations-title": "Donations",
+        "crypto-title": "Cryptocurrency",
+        "bank-desc": "The most convenient way with no commission",
+        "donations-desc": "If you want your message to appear on the stream",
+        "crypto-desc": "For connoisseurs of anonymity and modern technology",
+        "copy-to-clipboard": "Click to copy",
+        "copied-feedback": "Copied!",
+        "share-title": "Link Builder",
+        "share-desc": "Select sections to display:",
+        "share-copy": "Copy",
+        "share-copied": "Copied!",
+        "share-button-aria": "Share",
+        "share-header": "Header",
+        "share-alt-header": "Alternative Header",
+        "share-subtitle": "Subtitle",
+        "share-full-name": "Full Name",
+        "optimal": "Optimal",
+        "lts": "Long Term Support"
+    }
+};
