+++ style.css
@@ -300,11 +300,10 @@ section:last-child { margin-bottom: 0; }
 .skill-cpp { background: #00599c; color: white; --focus-color: #00599c; }
 .skill-java { background: #5382a1; color: #f0f0f0; --focus-color: #5382a1; }
 .skill-python { background: #3776ab; color: #f0f0f0; --focus-color: #3776ab; }
-.skill-sql { background: #00a9ff; color: #005073; --focus-color: #00a9ff; }
-.skill-unity { background: #0e1120; color: white; --focus-color: #888888; }
+.skill-sql { background: #007bff; color: white; --focus-color: #007bff; }
+.skill-fl { background: #e67e22; color: white; --focus-color: #e67e22; }
 .skill-unreal { background: #000000; color: white; --focus-color: #888888; }
 .skill-android { background: #3ddc84; color: #0c3b2e; --focus-color: #3ddc84; }
-.skill-fl { background: #ff9800; color: #803d00; --focus-color: #ff9800; }
 .skill-davinci { background: #0a0a2a; color: white; --focus-color: #4b4b6a; }
 .skill-ai { background: #8e44ad; color: white; --focus-color: #8e44ad; }
 
@@ -896,7 +895,8 @@ a { text-decoration: none; color: inherit; }
 }
 
 .tbank-commission { color: var(--secondary); }
-.donatty-commission { color: var(--warning); }
+.donatty-commission { color: #b45309; }
+[data-theme="dark"] .donatty-commission { color: var(--warning); }
 .donationalerts-commission { color: var(--danger); }
 .crypto-commission {
     color: var(--text-secondary);
